refactor(optimizer): Split subquery_to_join.rs into smaller modules
Split the 1166-line subquery_to_join.rs into logical modules for better
maintainability:

- mod.rs (197 lines): Main orchestration and public API
- in_clause.rs (167 lines): IN/NOT IN → SEMI/ANTI join transformations
- exists.rs (127 lines): EXISTS/NOT EXISTS → SEMI/ANTI join transformations
- helpers.rs (211 lines): Table name detection and column rewriting utilities
- tests.rs (540 lines): All unit tests

All existing tests pass. No public API changes.

Closes #3055
